+++ src/main/java/com/group1/shop_runner/repository/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Integer>,
+public interface OrderRepository extends JpaRepository<Order, Long>,
         JpaSpecificationExecutor<Order> {
     List<Order> findByUserId(Long userId);
 }
